Update README.md
This is a fully functional e-commerce website developed using HTML, CSS, and JavaScript. It features a user-friendly interface with product listings, a shopping cart, and checkout functionalities. The website is designed to provide a seamless shopping experience, allowing users to browse products, add them to the cart, and proceed to checkout. This project showcases my skills in web development and front-end design